Migrate F/A-18C to new module type
