+++ Scripts/DCS-BIOS/doc/Addresses.h
@@ -7132,12 +7132,10 @@
 #define FA_18C_hornet_CMSD_JET_SEL_L_AM 0x74D4, 0x8000
 #define FA_18C_hornet_COCKKPIT_LIGHT_MODE_SW 0x74C8, 0x0600, 9
 #define FA_18C_hornet_COMM1_ANT_SELECT_SW 0x74C0, 0x3000, 12
-#define FA_18C_hornet_COMM1_CHANNEL_NUMERIC 0x7404, 0xFFFF, 0
-#define FA_18C_hornet_COMM1_CHANNEL_NUMERIC_A 0x7404
+#define FA_18C_hornet_COMM1_CHANNEL_NUMERIC 0x7404, 0x001F, 0
 #define FA_18C_hornet_COMM1_FREQ 0x7400, 0xFFFF, 0
 #define FA_18C_hornet_COMM1_FREQ_A 0x7400
-#define FA_18C_hornet_COMM2_CHANNEL_NUMERIC 0x7406, 0xFFFF, 0
-#define FA_18C_hornet_COMM2_CHANNEL_NUMERIC_A 0x7406
+#define FA_18C_hornet_COMM2_CHANNEL_NUMERIC 0x7406, 0x001F, 0
 #define FA_18C_hornet_COMM2_FREQ 0x7402, 0xFFFF, 0
 #define FA_18C_hornet_COMM2_FREQ_A 0x7402
 #define FA_18C_hornet_COM_AUX 0x7538, 0xFFFF, 0
